requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
boto3==1.17.83
botocore==1.20.83
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
coverage==5.5
cryptography==3.4.7
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-environ==0.4.5
django-heroku==0.3.1
django-rest-knox==4.1.0
django-storages==1.11.1
djangorestframework==3.12.4
filelock==3.0.12
google-api-core==1.28.0
google-auth==1.30.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
isort==5.8.0
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mux-python @ git+https://github.com/muxinc/mux-python.git@c2df765035d457dd98ae69c3466e48f22964c43a
oauthlib==3.1.0
packaging==20.9
Pillow==8.2.0
pipenv==2020.11.15
protobuf==3.17.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pylint==2.8.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.4.7
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
